Add explicit integer casts
WebKit uses the -Wshorten-64-to-32 flag which warns on these cases.

Bug: 3439

+++ util/osx/OSXWindow.mm
@@ -622,7 +622,7 @@ OSXWindow::~OSXWindow()
     destroy();
 }
 
-bool OSXWindow::initialize(const std::string &name, size_t width, size_t height)
+bool OSXWindow::initialize(const std::string &name, int width, int height)
 {
     if (!InitializeAppKit())
     {
